major refactor of blockchain class
