+++ src/server/block_store.cpp
@@ -23,7 +23,7 @@ void BlockStore::setBlockCount(size_t count) {
     if(!status.ok()) throw std::runtime_error("Could not write block count to DB : " + status.ToString());
 }
 
-size_t BlockStore::getBlockCount() {
+size_t BlockStore::getBlockCount() const {
     string countKey = BLOCK_COUNT_KEY;
     leveldb::Slice key = leveldb::Slice(countKey);
     string value;
@@ -44,7 +44,7 @@ void BlockStore::setTotalWork(Bigint count) {
     if(!status.ok()) throw std::runtime_error("Could not write block count to DB : " + status.ToString());
 }
 
-Bigint BlockStore::getTotalWork() {
+Bigint BlockStore::getTotalWork() const {
     string countKey = TOTAL_WORK_KEY;
     leveldb::Slice key = leveldb::Slice(countKey);
     string value;
@@ -54,7 +54,7 @@ Bigint BlockStore::getTotalWork() {
     return b;
 }
 
-bool BlockStore::hasBlockCount() {
+bool BlockStore::hasBlockCount() const {
     string countKey = BLOCK_COUNT_KEY;
     leveldb::Slice key = leveldb::Slice(countKey);
     string value;
@@ -63,14 +63,14 @@ bool BlockStore::hasBlockCount() {
     return (status.ok());
 }
 
-bool BlockStore::hasBlock(uint32_t blockId) {
+bool BlockStore::hasBlock(uint32_t blockId) const {
     leveldb::Slice key = leveldb::Slice((const char*) &blockId, sizeof(uint32_t));
     string value;
     leveldb::Status status = db->Get(leveldb::ReadOptions(),key, &value);
     return (status.ok());
 }
 
-BlockHeader BlockStore::getBlockHeader(uint32_t blockId) {
+BlockHeader BlockStore::getBlockHeader(uint32_t blockId) const {
     leveldb::Slice key = leveldb::Slice((const char*) &blockId, sizeof(uint32_t));
     string valueStr;
     leveldb::Status status = db->Get(leveldb::ReadOptions(),key, &valueStr);
@@ -81,7 +81,7 @@ BlockHeader BlockStore::getBlockHeader(uint32_t blockId) {
     return value;
 }
 
-vector<TransactionInfo> BlockStore::getBlockTransactions(BlockHeader& block) {
+vector<TransactionInfo> BlockStore::getBlockTransactions(BlockHeader& block) const {
     vector<TransactionInfo> transactions;
     for(int i = 0; i < block.numTransactions; i++) {
         int idx = i;
@@ -99,7 +99,7 @@ vector<TransactionInfo> BlockStore::getBlockTransactions(BlockHeader& block) {
     return std::move(transactions);
 }
 
-std::pair<uint8_t*, size_t> BlockStore::getRawData(uint32_t blockId) {
+std::pair<uint8_t*, size_t> BlockStore::getRawData(uint32_t blockId) const {
     BlockHeader block = this->getBlockHeader(blockId);
     size_t numBytes = BLOCKHEADER_BUFFER_SIZE + (TRANSACTIONINFO_BUFFER_SIZE * block.numTransactions);
     char* buffer = (char*)malloc(numBytes);
@@ -122,7 +122,7 @@ std::pair<uint8_t*, size_t> BlockStore::getRawData(uint32_t blockId) {
     return std::pair<uint8_t*, size_t>((uint8_t*)buffer, numBytes);
 }
 
-Block BlockStore::getBlock(uint32_t blockId) {
+Block BlockStore::getBlock(uint32_t blockId) const {
     BlockHeader block = this->getBlockHeader(blockId);
     vector<TransactionInfo> transactionInfo = this->getBlockTransactions(block);
     vector<Transaction> transactions;
@@ -133,7 +133,7 @@ Block BlockStore::getBlock(uint32_t blockId) {
     return ret;
 }
 
-vector<SHA256Hash> BlockStore::getTransactionsForWallet(PublicWalletAddress& wallet) {
+vector<SHA256Hash> BlockStore::getTransactionsForWallet(PublicWalletAddress& wallet) const {
     struct {
         uint8_t addr[25];
         uint8_t txId[32];
