+++ src/server/block_store.hpp
@@ -8,19 +8,19 @@
 class BlockStore : public DataStore {
     public:
         BlockStore();
-        bool hasBlock(uint32_t blockId);
-        Block getBlock(uint32_t blockId);
-        std::pair<uint8_t*, size_t> getRawData(uint32_t blockId);
-        BlockHeader getBlockHeader(uint32_t blockId);
+        bool hasBlock(uint32_t blockId) const;
+        Block getBlock(uint32_t blockId) const;
+        std::pair<uint8_t*, size_t> getRawData(uint32_t blockId) const;
+        BlockHeader getBlockHeader(uint32_t blockId) const;
         void setBlock(Block& b);
         void setBlockCount(size_t count);
-        size_t getBlockCount();
+        size_t getBlockCount() const;
         void setTotalWork(Bigint work);
-        Bigint getTotalWork();
-        bool hasBlockCount();
+        Bigint getTotalWork() const;
+        bool hasBlockCount() const;
 
-        vector<SHA256Hash> getTransactionsForWallet(PublicWalletAddress& wallet);
+        vector<SHA256Hash> getTransactionsForWallet(PublicWalletAddress& wallet) const;
         void removeBlockWalletTransactions(Block& block);
     protected:
-        vector<TransactionInfo> getBlockTransactions(BlockHeader& block);
+        vector<TransactionInfo> getBlockTransactions(BlockHeader& block) const;
 };
